First test for Interventions controller
Using Moq to mock DbSet and other dependencies. New test method needs to be refactored, for it requires a great deal of configuration and mocking which should exist in other methods.

New tests should be far easier to develop than this one.

Changed the behaviour of "Interventions [] operator" to return the nth intervention in the set, not the intervention with the given GUID.